Rename the widget to SimpleScanViewer
