+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/simplescanviewer/rectangle_controller.py
