+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/simplescanviewer/rectangle_plot.py
