+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/simplescanviewer/rectangle_selection.py
